Add DBProvider.kt
