+++ app/src/main/java/com/example/newsandroid/database/Room.kt
@@ -1,6 +1,5 @@
 package com.example.newsandroid.database
 
-import android.content.Context
 import androidx.lifecycle.LiveData
 import androidx.room.*
 
@@ -16,18 +15,4 @@ interface NewsDao {
 @Database(entities = [DatabaseNews::class], version = 1)
 abstract class NewsDatabase: RoomDatabase() {
     abstract val newsDao: NewsDao
-}
-
-private lateinit var INSTANCE: NewsDatabase
-private val dbName = "news"
-
-fun getDatabase(context: Context): NewsDatabase {
-    synchronized(NewsDatabase::class.java) {
-        if (!::INSTANCE.isInitialized) {
-            INSTANCE = Room.databaseBuilder(context.applicationContext,
-                NewsDatabase::class.java,
-                dbName).build()
-        }
-    }
-    return INSTANCE
 }
