+++ app/src/main/java/com/example/newsandroid/ui/topheadlines/TopHeadlinesViewModel.kt
@@ -5,7 +5,7 @@ import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.example.newsandroid.database.getDatabase
+import com.example.newsandroid.database.DBProvider
 import com.example.newsandroid.repository.NewsRepository
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -17,7 +17,7 @@ enum class NewsApiStatus { LOADING, ERROR, DONE }
 
 class TopHeadlinesViewModel(application: Application) : ViewModel() {
 
-    private val newsRepository = NewsRepository(getDatabase(application))
+    private val newsRepository = NewsRepository(DBProvider.getDatabase(application))
 
     private val _status = MutableLiveData<NewsApiStatus>()
